refactor: Mark any case critical inline; publish cases via test-kit; rename streamable-http dimension to 2025-11-25
Replaces the hand-duplicated Scenario/scenarios.ts sharing mechanism with a
single Case model: every integration case is now a plain { name, critical,
run, ... } object defined once in src/test_kit/cases/*.cases.ts (published
behind the ./test-kit export). Marking a case critical is a one-line edit on
its own definition — no second array or file to keep in sync. This repo's own
suite.ts registers every case via registerCases(name, cases, ctx); internal
will register the same arrays with criticalOnly: true and pick up every
current and future critical case automatically.

Folds in and supersedes the previous commit's Scenario-based
scenarios.ts/registerScenarios mechanism entirely — that intermediate step
never shipped; every line it added is deleted here in favor of the Case
model below.

- src/test_kit/types.ts: Scenario -> Case, ScenarioCtx -> CaseCtx, adds
  hasTasksSupport/skipIf/retry so per-case transport gating and known-flaky
  retries move with the case instead of living in a separate registration call.
- src/test_kit/register.ts: registerScenarios -> registerCases, honors skipIf.
- src/test_kit/helpers.ts (new): merges tests/integration/cases/shared.ts,
  tests/const.ts, tests/integration/utils/task_waits.ts, and
  tests/helpers.ts's asLegacyClient into one published module, plus the new
  withClient(options, testFn) helper every simple case's run is built from.
- src/test_kit/cases/*.cases.ts (new): registration/tools/actors/apps/tasks/
  storage/payments, migrated from tests/integration/cases/*.cases.ts. All 9
  tasks cases and the 2 pre-existing shared scenarios are now critical: true,
  folded directly into their group instead of a separate registerScenarios
  call — this also removes one accidental duplicate run of "should list all
  default tools and Actors" (previously registered twice: once as its own
  itc() case, once via the shared-scenario describe block).
- payments.cases.ts: widened from streamable-http-only to also run on the
  2026-07-28 v2 client (plain listTools/callTool cases), except the one
  task-mode case that needs the v1-only experimental.tasks API.
- storage: the 5 self-contained cases move to src/test_kit/cases/storage.
  cases.ts; the 13 cases sharing one beforeAll-seeded Actor run don't fit the
  flat Case model (would mean paying for 13x redundant Actor runs) and stay
  local in tests/integration/cases/storage_grouped.cases.ts — not published,
  not critical-eligible.
- Transport dimension literal 'streamable-http' -> '2025-11-25', symmetric
  with '2026-07-28' and the 2025_11_25/2026_07_28 naming already used by the
  conformance-test config. The MCP protocol transport-type string
  ('streamable-http' in server_card.ts/server.json) is untouched — only the
  test-suite dimension label changes.
- oxlint.config.ts: scoped override disabling vitest/no-standalone-expect and
  vitest/expect-expect for src/test_kit/** — withClient()'s indirection hides
  expect() calls from the linter's static it()-nesting check even though they
  run inside a real it() at runtime (see register.ts).
- Deletes tests/integration/cases/shared.ts, shared_scenarios.ts, and
  src/test_kit/scenarios.ts (superseded).

Verified: type-check, lint, format, unit tests (91 files/1267 tests), build
(dist/test_kit/cases/*.js present), check:agents, and an AST-based diff of
every case's title against the pre-migration source — 117 unique tests
collected per entry file (was 118, exactly the one deduplicated case), 353
total across all 4 integration entry files (down from 356 by the same 3x1).

test: Split the default-listing case into explicit telemetry on/off cases, mark critical

The old single case asserted telemetry off (report-problem absent) via the
suite's own default, never setting telemetry explicitly. internal's own
tool-loading.cases.ts carried a near-duplicate asserting the opposite
(report-problem present, since its hosted deploy runs telemetry on by
default) - same scenario, diverging only because of which suite's default
happened to apply, not a real behavioral difference worth two separate
hand-maintained tests in two repos.

Now: two cases, each forcing telemetry via withClient's own option instead
of relying on any suite/environment default, both critical so internal gets
both automatically. Each asserts the full tools list, not just
report-problem's presence - confirms toggling telemetry changes only that
one tool, nothing else shifts.

Also marks three more cases critical, reviewed for coverage gaps:
- x402 payment-required error without signature, and skyfire-pay-id
  injection - payment gating is security/revenue relevant, and neither
  was shared with internal before.
- 'should call Actor directly with required input' - cheapest full
  round-trip smoke test (auth + baseUrl + real Actor run), same class as
  the two wrong-environment bugs already found in this branch.
