+++ AGENTS.md
@@ -47,7 +47,7 @@ After `pnpm run build`, run `mcpc` (no args) to check sessions: if `@stdio` (def
 - **Integration tests**: `pnpm run test:integration` (needs build + `APIFY_TOKEN`, humans only).
 - **Conformance tests**: `pnpm run test:conformance` (needs `APIFY_TOKEN`) builds and runs both protocol eras locally. `_integration_tests.yaml` calls `_conformance_tests.yaml` for the same `2026-07-28` and `2025-11-25` coverage in CI. Both eras always run; the command exits with the first non-zero code. Scenarios excluded from each era's gate are listed with reasons in `scripts/conformance_expected_failures_2026_07_28.yaml` and `scripts/conformance_expected_failures_2025_11_25.yaml`.
 - **Package manager**: this repo uses **pnpm 11+**. `devEngines.packageManager` is pinned with `onFail: "error"`, so npm / yarn refuse to run inside the checkout — use `pnpm install` only.
-- `tests/integration/suite.ts` wires the stdio, streamable-http and `2026-07-28` (stateless HTTP) transport dimensions into one shared suite. Add new integration cases to the matching capability file under `tests/integration/cases/*.cases.ts` (registration, tools, actors, apps, tasks, storage, payments) — each exports a `register*Cases(ctx)` called from `suite.ts`. Shared assertion helpers live in `tests/integration/cases/shared.ts`. A small subset shared with `apify-mcp-server-internal` (via `@apify/actors-mcp-server/test-kit`, `critical: true`) lives in `tests/integration/cases/shared_scenarios.ts` — see `src/test_kit/`.
+- `tests/integration/suite.ts` wires the `stdio`, `2025-11-25` (streamable HTTP) and `2026-07-28` (stateless HTTP) transport dimensions into one shared suite. Add new integration cases to the matching capability array in `src/test_kit/cases/*.cases.ts` (registration, tools, actors, apps, tasks, storage, payments) — each is a plain `Case[]` (`{ name, critical, run, ... }`), registered via `registerCases(name, cases, ctx)`. Mark a case `critical: true` to share it with `apify-mcp-server-internal` via `@apify/actors-mcp-server/test-kit` — no separate file or array to keep in sync. Shared assertion helpers live in `src/test_kit/helpers.ts`. One group (13 storage cases sharing a `beforeAll`-seeded Actor run) doesn't fit the flat model and stays local in `tests/integration/cases/storage_grouped.cases.ts`.
 - Follow existing test patterns (names, structure) — check neighboring files.
 - **Test naming**: `describe('fnName()')`, plain-verb `it()` names (no `should` prefix). Group with nested `describe()` per method when a factory/class exposes several.
